Remove mozart_quiesce() & mozart_rock_and_roll() from mozart_add_to_playlist()
mozart_quiesce() resets the playlist. So calling this each time an
enrty has been added to the playlist is undesirable.

Calling mozart_rock_and_roll() each time an entry is added to the
playlist is also undesirable as this gets the track index out of
whack, but does by chance play the right track,

mozart_rock_and_roll() should be called from the application. It is
still called within libmozart on an "about-to-finish" signal.