Don't try to remove the active playlist.
Fix a bug in mozart_remove_playlist() where we where passing
active_playlist to the find_list function to find the node of
the list we wanted to remove. We should have been passing the
specified playlist.

Also add a check to make sure we don't try to remove the active
playlist.