Fix transitioning from an internet stream to a local file.
When sending an "about-to-finish" signal in e.g
mozart_switch_playlist while an internet radio stream is
playing, we were setting the element stste to READY.

When going from an internet radio station to a local file
we were never getting out the of READY state. Setting the
state to NULL prior to sending the "about-to-finish" signal
seems to fix this.

This also needed doing in the mozart_{next,prev}_track functions.