Create a new debugging infrastructure.
We now have a d_printf macro that is basically just a wrapper
around printf.

You call it just like printf, but with a new first argument that
is the desired debug level. e.g

	d_printf(2, "test %s: Some debug info.\n", __FUNCTION__);

This will only print the message if the debug level (set through
the LIBMOZART_DEBUG environment variable) is set to 2 or more.

This is somewhat inspired from printk in the Linux kernel.