Make the active_playlist_index mean what it says.
This is now initialised to -1 and incremented by 1 first thing now in
mozart_rock_and_roll and all code that relied on active_playlist_index
has been updated accordingly.

This means that active_playlist_index is now the position in the playlist
array during the playing of that track. i.e it has the range
0..nr_tracks - 1. If it equals nr_tracks that means we've hit the end of
the playlist and can act accordingly.