Simplify the return value calculation in mozart_get_stream_duration_sec()
We only want the whole number of seconds for the duration. An int / int
will give an int with the fraction part dropped (floor).

This change doesn't alter the value that is now being returned.

As a side note, simply dropping the fraction part of the track
duration seems to be in line with the likes of Taglib and Rhythmbox.

e.g a track that is 4m:43.826s long is regarded as being 4m43s