Simplyfy seeking.
Combine the long and short seeks into one seek, that seeks 60
seconds forwards and 10 seconds backwards.

The reason for the length difference is to make it easier to get
to a specific point.

If you have just heard something you want to hear again, going back
a minute might be too much, but going forwards a muinute from there
likely isn't going to help.