Fix a problem where getting the stream duration could fail.
There are two functions

	mozart_play_index_at_pos()
	mozart_play_uri_at_pos()

Which allow you to jump to a specific track/uri and start playing
at a specific position within the stream.

However it seems after a call to these functions subsequent calls
to gst_element_query_duration() on that stream _could_ fail.

The fix seems to be to set the stream to state PAUSED and sleep a
bit before setting the stream to PLAYING.