libmozart: Fix a miscalculation of minutes duration
In mozart_convert_seconds_to_hms when we had a track longer
than 3599 seconds we would calculate the wrong value for the
minutes e.g

Given a track of length 5325 seconds (1h, 28m, 45s), we would
simply divide that by 60 to get 88 minutes.

Instead we need to divide the previously calculated value of
seconds (secs - (*hours * 3600)) by 60, which in this case
gives, 1725 / 60 = 28. Correct.